+++ src/enums/button-label.enum.ts
@@ -1,0 +1,4 @@
+export enum ButtonLabel {
+  Code = "Code",
+  SeeMore = "See more",
+}
