--- src/enums/button-selector.enum.ts
@@ -1,4 +1,0 @@
-export enum ButtonLabel {
-    Code = "Code",
-    SeeMore = "See more",
-}
